Fix spectator crashing when converting mania replay frames
